requirements.txt
GitPython~=3.1.0
PyYAML~=5.3.0
semver~=2.10.0
node-semver~=0.8.0